requirements.txt
PySide6
google-genai
